+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Changed the signatures of `isEqualToWithGivenProperties` and `isEqualToIgnoringGivenProperties` to 
+be able to take nullable properties.
+
 ## [0.20] 2019-09-13
 
 ### Added